gaits: Fix a problem with the action list in three ways:  1. Stop sorting the action list.  We're building it up in the correct     order to begin with.  Sorting only makes it possible to get out of     order both because python's sort isn't stable, and from floating     point errors.  2. Execute all actions which are ready without calling     _noaction_advance_phase, rather than making a zero call to     _noaction_advance_phase in between each ready action.  This way,     _noaction_advance_phase can assume that it will only be invoked in     between atomic action groups.  3. In advance_phase, never mod the phase by 1.0, instead rely on     hitting the end of the action list where we decrease it by 1.0.     This means that the phase can't get out of sync with the action list,     skipping final elements of it.
Add unit tests which cover a couple of commands, verifying a bunch of
properties of all of the states and motion.

Add the _mm suffix to the config's body_z_offset to be consistent with
the other parameters.

When returning the idle state, return all the stance legs in the world
frame so that we can just assert that stance legs are always in the
world frame.

When copying state objects, copy the phase as well.