Improve the jumping heuristics
Now, we reliably switch to landing mode shortly after touching the
ground, since we just look for displacement instead of velocity.
Also, continually calculate a trajectory when landing so we
semi-accurately approach the bottom of the jump.  This still doesn't
work all that well yet, so impose a minimum downward velocity also.