Fix conflation of M frame
We were in some places assuming the M frame was locked to the gyro
heading, and other places assuming it was locked to the body heading.
Make two separate frames with those two conventions and use them where
appropriate.

The net result was that the CoM was accounted for incorrectly, in a
way that would vary depending upon which direction the robot was
pointed!

While we are at it, make tplot2 able to view from any of the relevant
frames, and make it able to show commanded support regions, to make
debugging these types of problems easier.